+++ include/elio/coro/task.hpp
@@ -45,6 +45,7 @@ struct final_awaiter {
     template<typename Promise>
     [[nodiscard]] std::coroutine_handle<> await_suspend(std::coroutine_handle<Promise> h) noexcept {
         auto continuation = h.promise().continuation_;
+        h.promise().unlink_parent_cancellation();
         h.promise().leave_frame_context();
         if (continuation) {
             return continuation;
@@ -122,7 +123,7 @@ struct task_access {
 
 struct join_state_base {
     join_state_base()
-        : execution_context_(std::make_shared<task_execution_context>()) {}
+        : execution_context_(detail::make_task_execution_context()) {}
 
     explicit join_state_base(
         std::shared_ptr<task_execution_context> execution_context)